requirements.txt
streamlit==1.35.0
ultralytics
opencv-python-headless
Pillow
numpy
pandas==2.2.2
altair==5.3.0
protobuf==3.20.3